Add abstract ValueStringBase class
